+++ changelog.md
@@ -7,6 +7,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Fix printing of additional non-grouped default options in help screen.
+
 ## [1.1.0 (2021-10-28)](https://github.com/kdeldycke/click-extra/compare/v1.0.1...v1.1.0)
 
 - Add a `--config`/`-C` option to load CLI configuration from a TOML file.